coriandrum/requirements.txt
django==1.11.6
pillow==4.3.0
djangorestframework==3.7.1
requests==2.18.4